+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] - 2018-04-18
+### Improved
+- Sonos queue is cleared before queuing music.
+
+### Fixed
+- `repeat` and `shuffle` settings for Sonos Actions actually work now.
+
 ## [1.1.0] - 2018-04-17
 ### Added
 - `room` parameter to general config. [Documentation](/docs/install.md#room).